+++ tests/conftest.py
@@ -1,3 +1,5 @@
+import ssl
+
 import pytest
 import trustme
 from cryptography.hazmat.backends import default_backend
@@ -10,7 +12,7 @@ def tls_certificate_authority() -> trustme.CA:
 
 
 @pytest.fixture
-def tls_certificate(tls_certificate_authority):
+def tls_certificate(tls_certificate_authority: trustme.CA) -> trustme.LeafCert:
     return tls_certificate_authority.issue_server_cert(
         "localhost",
         "127.0.0.1",
@@ -19,13 +21,13 @@ def tls_certificate(tls_certificate_authority):
 
 
 @pytest.fixture
-def tls_ca_certificate_pem_path(tls_certificate_authority):
+def tls_ca_certificate_pem_path(tls_certificate_authority: trustme.CA):
     with tls_certificate_authority.cert_pem.tempfile() as ca_cert_pem:
         yield ca_cert_pem
 
 
 @pytest.fixture
-def tls_ca_certificate_private_key_path(tls_certificate_authority):
+def tls_ca_certificate_private_key_path(tls_certificate_authority: trustme.CA):
     with tls_certificate_authority.private_key_pem.tempfile() as private_key:
         yield private_key
 
@@ -47,6 +49,13 @@ def tls_ca_certificate_private_key_encrypted_path(tls_certificate_authority):
 
 
 @pytest.fixture
-def tls_certificate_pem_path(tls_certificate):
+def tls_certificate_pem_path(tls_certificate: trustme.LeafCert):
     with tls_certificate.private_key_and_cert_chain_pem.tempfile() as cert_pem:
         yield cert_pem
+
+
+@pytest.fixture
+def tls_ca_ssl_context(tls_certificate: trustme.LeafCert) -> ssl.SSLContext:
+    ssl_ctx = ssl.SSLContext()
+    tls_certificate.configure_cert(ssl_ctx)
+    return ssl_ctx