+++ tests/middleware/test_proxy_headers.py
@@ -1,4 +1,6 @@
-from tests.client import TestClient
+import httpx
+import pytest
+
 from tests.response import Response
 from uvicorn.middleware.proxy_headers import ProxyHeadersMiddleware
 
@@ -11,31 +13,82 @@ async def app(scope, receive, send):
     await response(scope, receive, send)
 
 
-app = ProxyHeadersMiddleware(app, trusted_hosts="*")
+@pytest.mark.asyncio
+@pytest.mark.parametrize(
+    ("trusted_hosts", "response_text"),
+    [
+        # always trust
+        ("*", "Remote: https://1.2.3.4:0"),
+        # trusted proxy
+        ("127.0.0.1", "Remote: https://1.2.3.4:0"),
+        (["127.0.0.1"], "Remote: https://1.2.3.4:0"),
+        # trusted proxy list
+        (["127.0.0.1", "10.0.0.1"], "Remote: https://1.2.3.4:0"),
+        ("127.0.0.1, 10.0.0.1", "Remote: https://1.2.3.4:0"),
+        # request from untrusted proxy
+        ("192.168.0.1", "Remote: http://127.0.0.1:123"),
+    ],
+)
+async def test_proxy_headers_trusted_hosts(trusted_hosts, response_text):
+    app_with_middleware = ProxyHeadersMiddleware(app, trusted_hosts=trusted_hosts)
+    async with httpx.AsyncClient(
+        app=app_with_middleware, base_url="http://testserver"
+    ) as client:
+        headers = {"X-Forwarded-Proto": "https", "X-Forwarded-For": "1.2.3.4"}
+        response = await client.get("/", headers=headers)
 
-
-def test_proxy_headers():
-    client = TestClient(app)
-    headers = {"X-Forwarded-Proto": "https", "X-Forwarded-For": "1.2.3.4"}
-    response = client.get("/", headers=headers)
     assert response.status_code == 200
-    assert response.text == "Remote: https://1.2.3.4:0"
+    assert response.text == response_text
 
 
-def test_proxy_headers_no_port():
-    client = TestClient(app)
-    headers = {"X-Forwarded-Proto": "https", "X-Forwarded-For": "1.2.3.4"}
-    response = client.get("/", headers=headers)
+@pytest.mark.asyncio
+@pytest.mark.parametrize(
+    ("trusted_hosts", "response_text"),
+    [
+        # always trust
+        ("*", "Remote: https://1.2.3.4:0"),
+        # all proxies are trusted
+        (
+            ["127.0.0.1", "10.0.2.1", "192.168.0.2"],
+            "Remote: https://1.2.3.4:0",
+        ),
+        # order doesn't matter
+        (
+            ["10.0.2.1", "192.168.0.2", "127.0.0.1"],
+            "Remote: https://1.2.3.4:0",
+        ),
+        # should set first untrusted as remote address
+        (["192.168.0.2", "127.0.0.1"], "Remote: https://10.0.2.1:0"),
+    ],
+)
+async def test_proxy_headers_multiple_proxies(trusted_hosts, response_text):
+    app_with_middleware = ProxyHeadersMiddleware(app, trusted_hosts=trusted_hosts)
+    async with httpx.AsyncClient(
+        app=app_with_middleware, base_url="http://testserver"
+    ) as client:
+        headers = {
+            "X-Forwarded-Proto": "https",
+            "X-Forwarded-For": "1.2.3.4, 10.0.2.1, 192.168.0.2",
+        }
+        response = await client.get("/", headers=headers)
+
     assert response.status_code == 200
-    assert response.text == "Remote: https://1.2.3.4:0"
+    assert response.text == response_text
 
 
-def test_proxy_headers_invalid_x_forwarded_for():
-    client = TestClient(app)
-    headers = {
-        "X-Forwarded-Proto": "https",
-        "X-Forwarded-For": "\xf0\xfd\xfd\xfd, 1.2.3.4",
-    }
-    response = client.get("/", headers=headers)
+@pytest.mark.asyncio
+async def test_proxy_headers_invalid_x_forwarded_for():
+    app_with_middleware = ProxyHeadersMiddleware(app, trusted_hosts="*")
+    async with httpx.AsyncClient(
+        app=app_with_middleware, base_url="http://testserver"
+    ) as client:
+        headers = httpx.Headers(
+            {
+                "X-Forwarded-Proto": "https",
+                "X-Forwarded-For": "1.2.3.4, \xf0\xfd\xfd\xfd",
+            },
+            encoding="latin-1",
+        )
+        response = await client.get("/", headers=headers)
     assert response.status_code == 200
     assert response.text == "Remote: https://1.2.3.4:0"