+++ uvicorn/middleware/debug.py
@@ -1,52 +1,65 @@
 import html
 import traceback
+from typing import Union
+
+from asgiref.typing import (
+    ASGI3Application,
+    ASGIReceiveCallable,
+    ASGISendCallable,
+    ASGISendEvent,
+    HTTPResponseBodyEvent,
+    HTTPResponseStartEvent,
+    WWWScope,
+)
 
 
 class HTMLResponse:
-    def __init__(self, content, status_code):
+    def __init__(self, content: str, status_code: int):
         self.content = content
         self.status_code = status_code
 
-    async def __call__(self, scope, recieve, send):
-        await send(
-            {
-                "type": "http.response.start",
-                "status": self.status_code,
-                "headers": [[b"content-type", b"text/html; charset=utf-8"]],
-            }
-        )
-        await send(
-            {
-                "type": "http.response.body",
-                "body": self.content.encode("utf-8"),
-                "more_body": False,
-            }
-        )
+    async def __call__(
+        self, scope: WWWScope, receive: ASGIReceiveCallable, send: ASGISendCallable
+    ) -> None:
+        response_start: HTTPResponseStartEvent = {
+            "type": "http.response.start",
+            "status": self.status_code,
+            "headers": [(b"content-type", b"text/html; charset=utf-8")],
+        }
+        await send(response_start)
+
+        response_body: HTTPResponseBodyEvent = {
+            "type": "http.response.body",
+            "body": self.content.encode("utf-8"),
+            "more_body": False,
+        }
+        await send(response_body)
 
 
 class PlainTextResponse:
-    def __init__(self, content, status_code):
+    def __init__(self, content: str, status_code: int):
         self.content = content
         self.status_code = status_code
 
-    async def __call__(self, scope, recieve, send):
-        await send(
-            {
-                "type": "http.response.start",
-                "status": self.status_code,
-                "headers": [[b"content-type", b"text/plain; charset=utf-8"]],
-            }
-        )
-        await send(
-            {
-                "type": "http.response.body",
-                "body": self.content.encode("utf-8"),
-                "more_body": False,
-            }
-        )
+    async def __call__(
+        self, scope: WWWScope, receive: ASGIReceiveCallable, send: ASGISendCallable
+    ) -> None:
+        response_start: HTTPResponseStartEvent = {
+            "type": "http.response.start",
+            "status": self.status_code,
+            "headers": [(b"content-type", b"text/plain; charset=utf-8")],
+        }
+        await send(response_start)
+
+        response_body: HTTPResponseBodyEvent = {
+            "type": "http.response.body",
+            "body": self.content.encode("utf-8"),
+            "more_body": False,
+        }
+        await send(response_body)
 
 
-def get_accept_header(scope):
+def get_accept_header(scope: WWWScope) -> str:
     accept = "*/*"
 
     for key, value in scope.get("headers", []):
@@ -58,16 +71,18 @@ def get_accept_header(scope):
 
 
 class DebugMiddleware:
-    def __init__(self, app):
+    def __init__(self, app: ASGI3Application):
         self.app = app
 
-    async def __call__(self, scope, receive, send):
+    async def __call__(
+        self, scope: WWWScope, receive: ASGIReceiveCallable, send: ASGISendCallable
+    ) -> None:
         if scope["type"] != "http":
             return await self.app(scope, receive, send)
 
         response_started = False
 
-        async def inner_send(message):
+        async def inner_send(message: ASGISendEvent) -> None:
             nonlocal response_started, send
 
             if message["type"] == "http.response.start":
@@ -81,6 +96,7 @@ class DebugMiddleware:
                 raise exc from None
 
             accept = get_accept_header(scope)
+            response: Union[HTMLResponse, PlainTextResponse]
             if "text/html" in accept:
                 exc_html = html.escape(traceback.format_exc())
                 content = (