+++ uvicorn/server.py
@@ -1,5 +1,4 @@
 import asyncio
-import functools
 import logging
 import os
 import platform
@@ -9,8 +8,11 @@ import sys
 import threading
 import time
 from email.utils import formatdate
+from typing import List
 
 import click
+
+from ._handlers.http import handle_http
 
 HANDLED_SIGNALS = (
     signal.SIGINT,  # Unix signal 2. Sent by Ctrl+C.
@@ -76,7 +78,7 @@ class Server:
             extra={"color_message": color_message},
         )
 
-    async def startup(self, sockets=None):
+    async def startup(self, sockets: list = None) -> None:
         await self.lifespan.startup()
         if self.lifespan.should_exit:
             self.should_exit = True
@@ -84,17 +86,18 @@ class Server:
 
         config = self.config
 
-        create_protocol = functools.partial(
-            config.http_protocol_class, config=config, server_state=self.server_state
-        )
-
-        loop = asyncio.get_event_loop()
+        async def handler(
+            reader: asyncio.StreamReader, writer: asyncio.StreamWriter
+        ) -> None:
+            await handle_http(
+                reader, writer, server_state=self.server_state, config=config
+            )
 
         if sockets is not None:
             # Explicitly passed a list of open sockets.
             # We use this when the server is run from a Gunicorn worker.
 
-            def _share_socket(sock: socket) -> socket:
+            def _share_socket(sock: socket.SocketType) -> socket.SocketType:
                 # Windows requires the socket be explicitly shared across
                 # multiple workers (processes).
                 from socket import fromshare  # type: ignore
@@ -106,19 +109,20 @@ class Server:
             for sock in sockets:
                 if config.workers > 1 and platform.system() == "Windows":
                     sock = _share_socket(sock)
-                server = await loop.create_server(
-                    create_protocol, sock=sock, ssl=config.ssl, backlog=config.backlog
+                server = await asyncio.start_server(
+                    handler, sock=sock, ssl=config.ssl, backlog=config.backlog
                 )
                 self.servers.append(server)
+            listeners = sockets
 
         elif config.fd is not None:
             # Use an existing socket, from a file descriptor.
             sock = socket.fromfd(config.fd, socket.AF_UNIX, socket.SOCK_STREAM)
-            server = await loop.create_server(
-                create_protocol, sock=sock, ssl=config.ssl, backlog=config.backlog
+            server = await asyncio.start_server(
+                handler, sock=sock, ssl=config.ssl, backlog=config.backlog
             )
-            message = "Uvicorn running on socket %s (Press CTRL+C to quit)"
-            logger.info(message % str(sock.getsockname()))
+            assert server.sockets is not None  # mypy
+            listeners = server.sockets
             self.servers = [server]
 
         elif config.uds is not None:
@@ -126,24 +130,19 @@ class Server:
             uds_perms = 0o666
             if os.path.exists(config.uds):
                 uds_perms = os.stat(config.uds).st_mode
-            server = await loop.create_unix_server(
-                create_protocol, path=config.uds, ssl=config.ssl, backlog=config.backlog
+            server = await asyncio.start_unix_server(
+                handler, path=config.uds, ssl=config.ssl, backlog=config.backlog
             )
             os.chmod(config.uds, uds_perms)
-            message = "Uvicorn running on unix socket %s (Press CTRL+C to quit)"
-            logger.info(message % config.uds)
+            assert server.sockets is not None  # mypy
+            listeners = server.sockets
             self.servers = [server]
 
         else:
             # Standard case. Create a socket from a host/port pair.
-            addr_format = "%s://%s:%d"
-            if config.host and ":" in config.host:
-                # It's an IPv6 address.
-                addr_format = "%s://[%s]:%d"
-
             try:
-                server = await loop.create_server(
-                    create_protocol,
+                server = await asyncio.start_server(
+                    handler,
                     host=config.host,
                     port=config.port,
                     ssl=config.ssl,
@@ -153,9 +152,46 @@ class Server:
                 logger.error(exc)
                 await self.lifespan.shutdown()
                 sys.exit(1)
+
+            assert server.sockets is not None
+            listeners = server.sockets
+            self.servers = [server]
+
+        if sockets is None:
+            self._log_started_message(listeners)
+        else:
+            # We're most likely running multiple workers, so a message has already been
+            # logged by `config.bind_socket()`.
+            pass
+
+        self.started = True
+
+    def _log_started_message(self, listeners: List[socket.SocketType]) -> None:
+        config = self.config
+
+        if config.fd is not None:
+            sock = listeners[0]
+            logger.info(
+                "Uvicorn running on socket %s (Press CTRL+C to quit)",
+                sock.getsockname(),
+            )
+
+        elif config.uds is not None:
+            logger.info(
+                "Uvicorn running on unix socket %s (Press CTRL+C to quit)", config.uds
+            )
+
+        else:
+            addr_format = "%s://%s:%d"
+            host = "0.0.0.0" if config.host is None else config.host
+            if ":" in host:
+                # It's an IPv6 address.
+                addr_format = "%s://[%s]:%d"
+
             port = config.port
             if port == 0:
-                port = server.sockets[0].getsockname()[1]
+                port = listeners[0].getsockname()[1]
+
             protocol_name = "https" if config.ssl else "http"
             message = f"Uvicorn running on {addr_format} (Press CTRL+C to quit)"
             color_message = (
@@ -166,13 +202,10 @@ class Server:
             logger.info(
                 message,
                 protocol_name,
-                config.host,
+                host,
                 port,
                 extra={"color_message": color_message},
             )
-            self.servers = [server]
-
-        self.started = True
 
     async def main_loop(self):
         counter = 0
@@ -249,7 +282,7 @@ class Server:
         try:
             for sig in HANDLED_SIGNALS:
                 loop.add_signal_handler(sig, self.handle_exit, sig, None)
-        except NotImplementedError:
+        except NotImplementedError:  # pragma: no cover
             # Windows
             for sig in HANDLED_SIGNALS:
                 signal.signal(sig, self.handle_exit)